fix: add support for simultaneous stage wide starts
Each stage (starting at the stage registers) needs to start running when
run is called without depending on the prior stage to pass valid values
to the stage register. This would ensure that the data path is operational
even if valid data has not been passed to it. This functionality is necessary
to support stage level thread synchronization.